# server/Chatify.Web/swagger/v1/swagger.yaml
openapi: 3.0.1
info:
  title: Chatify Server
  version: v1
servers:
  - url: https://localhost:7139
paths:
...
  '/api/Messages/{groupId}':
    get:
      tags:
        - Messages
      parameters:
        - name: GroupId
          in: query
          style: form
          schema:
            type: string
            format: uuid
        - name: PageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: PagingCursor
          in: query
          style: form
          schema:
            type: string
        - name: groupId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatGroupMessageEntryCursorPagedApiResponse'
    post:
      tags:
        - Messages
      parameters:
        - name: groupId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                ChatGroupId:
                  type: string
                  format: uuid
                Content:
                  type: string
                Files:
                  type: array
                  items:
                    type: string
                    format: binary
                Metadata:
                  type: object
                  additionalProperties:
                    type: string
            encoding:
              ChatGroupId:
                style: form
              Content:
                style: form
              Files:
                style: form
              Metadata:
                style: form
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema: { }
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '202':
          description: Accepted
          content:
            application/json:
              schema: { }
  '/api/Messages/{messageId}/replies':
    get:
      tags:
        - Messages
      parameters:
        - name: MessageId
          in: query
          style: form
          schema:
            type: string
            format: uuid
        - name: PageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: PagingCursor
          in: query
          style: form
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageReplyCursorPagedApiResponse'
    post:
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendGroupChatMessageReplyRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/SendGroupChatMessageReplyRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SendGroupChatMessageReplyRequest'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectApiResponse'
  '/api/Messages/replies/{messageId}':
    delete:
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteGroupChatMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/DeleteGroupChatMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DeleteGroupChatMessageRequest'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '204':
          description: No Content
    put:
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditGroupChatMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/EditGroupChatMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EditGroupChatMessageRequest'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '202':
          description: Accepted
  '/api/Messages/{messageId}':
    put:
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditGroupChatMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/EditGroupChatMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EditGroupChatMessageRequest'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '202':
          description: Accepted
    delete:
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteGroupChatMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/DeleteGroupChatMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DeleteGroupChatMessageRequest'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '204':
          description: No Content
  '/pins/{messageId}':
    post:
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinChatGroupMessage'
          text/json:
            schema:
              $ref: '#/components/schemas/PinChatGroupMessage'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PinChatGroupMessage'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '204':
          description: No Content
  '/share/{messageId}':
    post:
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareMessage'
          text/json:
            schema:
              $ref: '#/components/schemas/ShareMessage'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ShareMessage'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitApiResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '202':
          description: Accepted
...
components:
  schemas:
...
    AddressFamily:
      enum:
        - Unknown
        - Unspecified
        - Unix
        - InterNetwork
        - ImpLink
        - Pup
        - Chaos
        - NS
        - Iso
        - Ecma
        - DataKit
        - Ccitt
        - Sna
        - DecNet
        - DataLink
        - Lat
        - HyperChannel
        - AppleTalk
        - NetBios
        - VoiceView
        - FireFox
        - Banyan
        - Atm
        - InterNetworkV6
        - Cluster
        - Ieee12844
        - Irda
        - NetworkDesigners
        - Max
        - Packet
        - ControllerAreaNetwork
      type: string
    ApiResponseStatus:
      enum:
        - Success
        - Failure
      type: string
...
    ChatGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creatorId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        about:
          type: string
          nullable: true
        picture:
          $ref: '#/components/schemas/Media'
        adminIds:
          uniqueItems: true
          type: array
          items:
            type: string
            format: uuid
          nullable: true
        admins:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/User'
          nullable: true
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
...
    ChatGroupMessageEntry:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/ChatMessage'
        forwardedMessage:
          $ref: '#/components/schemas/ChatMessage'
        senderInfo:
          $ref: '#/components/schemas/MessageSenderInfoEntry'
        repliersInfo:
          $ref: '#/components/schemas/MessageRepliersInfoEntry'
      additionalProperties: false
    ChatGroupMessageEntryCursorPagedApiResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ApiResponseStatus'
        errors:
          type: array
          items:
            type: string
          nullable: true
        message:
          type: string
          nullable: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ChatGroupMessageEntry'
          nullable: true
        timestamp:
          type: string
          format: date-time
      additionalProperties: false
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        chatGroupId:
          type: string
          format: uuid
        chatGroup:
          $ref: '#/components/schemas/ChatGroup'
        userId:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        content:
          type: string
          nullable: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Media'
          nullable: true
        reactionCounts:
          type: object
          additionalProperties:
            type: integer
            format: int64
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
      additionalProperties: false
...
    ChatMessageReply:
      type: object
      properties:
        id:
          type: string
          format: uuid
        chatGroupId:
          type: string
          format: uuid
        chatGroup:
          $ref: '#/components/schemas/ChatGroup'
        userId:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        content:
          type: string
          nullable: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Media'
          nullable: true
        reactionCounts:
          type: object
          additionalProperties:
            type: integer
            format: int64
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
        replyToId:
          type: string
          format: uuid
      additionalProperties: false
    ChatMessageReplyCursorPagedApiResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ApiResponseStatus'
        errors:
          type: array
          items:
            type: string
          nullable: true
        message:
          type: string
          nullable: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageReply'
          nullable: true
        timestamp:
          type: string
          format: date-time
      additionalProperties: false
    DeleteGroupChatMessageRequest:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
      additionalProperties: false
    EditGroupChatMessageRequest:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
        newContent:
          type: string
          nullable: true
      additionalProperties: false
...
    Email:
      type: object
      properties:
        value:
          type: string
          nullable: true
          readOnly: true
      additionalProperties: false
...
    IPAddress:
      type: object
      properties:
        addressFamily:
          $ref: '#/components/schemas/AddressFamily'
        scopeId:
          type: integer
          format: int64
        isIPv6Multicast:
          type: boolean
          readOnly: true
        isIPv6LinkLocal:
          type: boolean
          readOnly: true
        isIPv6SiteLocal:
          type: boolean
          readOnly: true
        isIPv6Teredo:
          type: boolean
          readOnly: true
        isIPv6UniqueLocal:
          type: boolean
          readOnly: true
        isIPv4MappedToIPv6:
          type: boolean
          readOnly: true
        address:
          type: integer
          format: int64
          deprecated: true
      additionalProperties: false
...
    Media:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mediaUrl:
          type: string
          nullable: true
        fileName:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
      additionalProperties: false
    MessageReplierInfoEntry:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
          nullable: true
        profilePictureUrl:
          type: string
          nullable: true
      additionalProperties: false
    MessageRepliersInfoEntry:
      type: object
      properties:
        total:
          type: integer
          format: int64
        lastUpdatedAt:
          type: string
          format: date-time
          nullable: true
        replierInfos:
          type: array
          items:
            $ref: '#/components/schemas/MessageReplierInfoEntry'
          nullable: true
      additionalProperties: false
    MessageSenderInfoEntry:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
          nullable: true
        profilePictureUrl:
          type: string
          nullable: true
      additionalProperties: false
...
    ObjectApiResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ApiResponseStatus'
        errors:
          type: array
          items:
            type: string
          nullable: true
        message:
          type: string
          nullable: true
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
      additionalProperties: false
    PhoneNumber:
      type: object
      properties:
        value:
          type: string
          nullable: true
          readOnly: true
      additionalProperties: false
    PinChatGroupMessage:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    SendGroupChatMessageReplyRequest:
      type: object
      properties:
        chatGroupId:
          type: string
          format: uuid
        replyToId:
          type: string
          format: uuid
        content:
          type: string
          nullable: true
        files:
          type: array
          items:
            type: string
            format: binary
          nullable: true
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
      additionalProperties: false
    ShareMessage:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        content:
          type: string
          nullable: true
      additionalProperties: false
...
    Unit:
      type: object
      additionalProperties: false
    UnitApiResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ApiResponseStatus'
        errors:
          type: array
          items:
            type: string
          nullable: true
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/Unit'
        timestamp:
          type: string
          format: date-time
      additionalProperties: false
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          nullable: true
        email:
          $ref: '#/components/schemas/Email'
        status:
          $ref: '#/components/schemas/UserStatus'
        roles:
          type: array
          items:
            type: string
          nullable: true
        phoneNumbers:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/PhoneNumber'
          nullable: true
        profilePicture:
          $ref: '#/components/schemas/Media'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        lastLogin:
          type: string
          format: date-time
        deviceIps:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/IPAddress'
          nullable: true
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
        displayName:
          type: string
          nullable: true
      additionalProperties: false
...
    UserStatus:
      enum:
        - Online
        - Away
        - Offline
      type: string